Null safety migration (#176)
This commit updates the flutter version and makes the code
conform to the newer nullsafety checks.

+++ test/lotus_test/private_key_test.dart
@@ -73,7 +73,7 @@ void main() {
 
     for (final arg in arguments) {
       final pk = BCHPrivateKey.fromBigInt(
-        arg[0],
+        arg[0] as BigInt,
       );
       expect(
         pk.toWIF(),
